requirements.txt
attrs>=21.0.0
matplotlib>=3.5.0
numpy>=1.21.0